[fix]  Fix issue with rendering incorrect data when x-axis aligned by `timestamp/epoch`
